docs: :lipstick: Move Table component to components/atoms and udate their reference
