+++ src/app/agents/page.tsx
@@ -3,7 +3,7 @@ import { useCallback, useEffect, useState } from "react";
 import { Agent, statusAgent } from "../../lib/models/agent";
 import { fetchAgents } from "../../lib/api/agents";
 import { useSearchParams, useRouter, usePathname } from "next/navigation";
-import { Table } from "@/components/Table";
+import { Table } from "@/components/atoms/Table";
 import { AgentFilter } from "../../lib/models/filter";
 import { createWebSocket } from "../../lib/ws/socket";
