+++ src/app/clients/page.tsx
@@ -3,7 +3,7 @@ import React, { useCallback, useEffect, useState } from "react";
 import { Client } from "../../lib/models/client";
 import { fetchClients } from "../../lib/api/clients";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
-import { Table } from "@/components/Table";
+import { Table } from "@/components/atoms/Table";
 import { filterOperations } from "../../lib/constants/operations";
 import { ClientFilter } from "../../lib/models/filter";
 import { createWebSocket } from "../../lib/ws/socket";
